# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: easyinterview API
  version: 1.0.0
  summary: HTTP contract for easyinterview P0 — JD onboarding, mock interview, evidence-based review.
  description: |
    Single source of truth for all easyinterview HTTP endpoints. Owned by spec
    `docs/spec/openapi-v1-contract/spec.md` (B2). Any change must follow the
    additive-only / breaking-change rules locked in spec §4.4 and ADR review flow.

    Auth path follows ADR-Q1: first-party session cookie minted by
    `GET /api/v1/auth/email/verify` after the magic-link challenge issued by
    `POST /api/v1/auth/email/start`. `Authorization: Bearer` is **not** a P0
    default scheme.
  contact:
    name: easyinterview engineering
    url: https://github.com/monshunter/easyinterview
servers:
  - url: /api/v1
    description: API v1 base path. All endpoints in this document are mounted under this prefix.
...
security:
  - sessionCookie: []
paths:
...
  /profiles/me:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: Get my candidate profile
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Candidate profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    patch:
      tags: [Profile]
      operationId: updateMyProfile
      summary: Update my candidate profile (Lite)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /profiles/me/experience-cards:
    get:
      tags: [Profile]
      operationId: listExperienceCards
      summary: List my experience cards (cursor-paginated)
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Paginated experience cards.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperienceCard'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      tags: [Profile]
      operationId: createExperienceCard
      summary: Create an experience card
      description: Side-effect endpoint per backend-profile spec D-5 — requires `Idempotency-Key`. Server forces `source_type='manual'` regardless of body.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExperienceCardRequest'
      responses:
        '201':
          description: Experience card created.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperienceCard'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /profiles/me/experience-cards/{cardId}:
    patch:
      tags: [Profile]
      operationId: updateExperienceCard
      summary: Update an experience card
      description: Side-effect endpoint per backend-profile spec D-5 — requires `Idempotency-Key`. Cross-user access returns 404 `RESOURCE_NOT_FOUND`.
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExperienceCardRequest'
      responses:
        '200':
          description: Updated experience card.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperienceCard'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ei_session
      description: |
        First-party session cookie minted by `GET /api/v1/auth/email/verify`.
        ADR-Q1 locks the cookie attributes (HttpOnly, SameSite=Lax, Secure, ~30d
        sliding renewal). Cookie value is opaque; the server-side `sessions`
        table is the source of truth. `Authorization: Bearer` is **not** an
        accepted P0 scheme — adding it requires a revision of ADR-Q1 and this
        spec.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Idempotency key per `B1 shared-conventions-codified §3.4` (24h TTL). Required by
        side-effect endpoints listed in spec D-6, except `POST /auth/email/start`
        (rate-limited per ADR-Q1) and `POST /practice/sessions/{sessionId}/events`
        (uses `clientEventId` instead).
      schema:
        type: string
        format: uuid
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Client-supplied correlation id; the server echoes it back via the X-Request-ID response header and `error.requestId`.
      schema:
        type: string
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: W3C Trace Context header.
      schema:
        type: string
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        UI-locale hint per `B1 shared-conventions-codified §3.5`. Affects display defaults only;
        does **not** override persisted business fields such as `targetLanguage`.
      schema:
        type: string
    XClientVersion:
      name: X-Client-Version
      in: header
      required: false
      description: Client build version reported by the frontend shell.
      schema:
        type: string
  headers:
    XRequestID:
      description: Server-confirmed correlation id; matches `error.requestId` on error responses.
      schema:
        type: string
  responses:
    ApiErrorResponse:
      description: Error envelope per `ApiError` schema.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    ApiErrorCode:
      type: string
      description: |
        Documented error codes synced from
        `shared/conventions.yaml#errors[].code`; B1 owns the literal set.
        P0 export exceptions such as `PRIVACY_EXPORT_NOT_AVAILABLE` and
        `RESUME_EXPORT_NOT_AVAILABLE` are included here so the OpenAPI
        contract is self-contained for 501 export responses.
      enum:
        - AUTH_UNAUTHORIZED
        - TARGET_IMPORT_FAILED
        - TARGET_JOB_NOT_FOUND
        - TARGET_IMPORT_SOURCE_INVALID
        - TARGET_IMPORT_SOURCE_UNAVAILABLE
        - TARGET_INVALID_STATE_TRANSITION
        - PRACTICE_SESSION_CONFLICT
        - PRACTICE_PLAN_NOT_FOUND
        - PRACTICE_SESSION_NOT_FOUND
        - REPORT_NOT_FOUND
        - REPORT_NOT_READY
        - DEBRIEF_NOT_FOUND
        - RESUME_EXPORT_NOT_AVAILABLE
        - RESUME_STRUCTURED_MASTER_ALREADY_EXISTS
        - VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - IDEMPOTENCY_KEY_MISMATCH
        - RATE_LIMITED
        - AI_PROVIDER_TIMEOUT
        - AI_OUTPUT_INVALID
        - AI_FALLBACK_EXHAUSTED
        - AI_UNSUPPORTED_CAPABILITY
        - AI_PROVIDER_CONFIG_INVALID
        - AI_PROVIDER_SECRET_MISSING
        - PRIVACY_EXPORT_NOT_AVAILABLE
    ApiError:
      type: object
      required: [code, message, requestId, retryable]
      description: |
        Canonical error object per B1 `structures.ApiError`. B2 wraps it in
        `ApiErrorResponse` for the wire body `{error: ...}`.
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
          description: Human-readable summary; safe to surface to operators (no secrets, per A4 redaction rules).
        requestId:
          type: string
          description: Request correlation id; equals the X-Request-ID response header for the same request.
        retryable:
          type: boolean
          description: Hint to the client whether retry is safe, as defined by B1 shared-conventions-codified.
        details:
          type: object
          description: Optional structured diagnostic. Must not contain provider, prompt, or secret material per A4 redaction rules.
          additionalProperties: true
    ApiErrorResponse:
      type: object
      required: [error]
      description: Wire error response envelope defined by B1 shared-conventions-codified.
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    PageInfo:
      type: object
      required: [pageSize, hasMore]
      description: Cursor pagination envelope per spec D-5. Mirrors B1 `structures.PageInfo`.
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page; `null` on the last page.
        pageSize:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        hasMore:
          type: boolean
...
    PaginatedEnvelope:
      type: object
      required: [items, pageInfo]
      description: |
        Base envelope for cursor-paginated list responses. Concrete
        `PaginatedFoo` schemas extend this via `allOf` and override `items`
        with a typed array, keeping the field order (`items`, `pageInfo`)
        consistent across endpoints.
      properties:
        items:
          type: array
          items: {}
        pageInfo:
          $ref: '#/components/schemas/PageInfo'

    # ----- Auth / runtime -----
...
    CandidateProfile:
      type: object
      required: [headline, yearsOfExperience, currentRole, preferredPracticeLanguage, uiLanguage, region]
      properties:
        headline:
          type: string
          nullable: true
        yearsOfExperience:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        currentRole:
          type: string
          nullable: true
        preferredPracticeLanguage:
          type: string
        uiLanguage:
          type: string
        region:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      description: All fields optional — only supplied fields are updated.
      properties:
        headline:
          type: string
        yearsOfExperience:
          type: integer
          format: int32
          minimum: 0
        currentRole:
          type: string
        preferredPracticeLanguage:
          type: string
        uiLanguage:
          type: string
        region:
          type: string
    ExperienceCard:
      type: object
      required: [id, title, companyName, situation, task, action, result, skills, language, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        companyName:
          type: string
        situation:
          type: string
        task:
          type: string
        action:
          type: string
        result:
          type: string
        skills:
          type: array
          items:
            type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExperienceCardRequest:
      type: object
      required: [title, companyName, situation, task, action, result, language]
      properties:
        title:
          type: string
        companyName:
          type: string
        situation:
          type: string
        task:
          type: string
        action:
          type: string
        result:
          type: string
        skills:
          type: array
          items:
            type: string
        language:
          type: string
    UpdateExperienceCardRequest:
      type: object
      properties:
        title:
          type: string
        companyName:
          type: string
        situation:
          type: string
        task:
          type: string
        action:
          type: string
        result:
          type: string
        skills:
          type: array
          items:
            type: string
        language:
          type: string
    PaginatedExperienceCard:
      allOf:
        - $ref: '#/components/schemas/PaginatedEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ExperienceCard'

    # ----- Resumes -----
...